refactor(scripts): split scale-children manifest from emission CLI
Move the 100-entry SHELL_MANIFEST ledger and capabilities-completeness
helpers into gen-scale-children-manifest.ts. Keep pure shell/index
emission, generate, and the package.json entrypoint in gen-scale-children.ts
so scale ledger edits no longer sit beside template and FS code.
